requirements.txt
flask
flask-cors
python-dotenv
openai
edge-tts
yt-dlp
youtube-transcript-api
requests
google-auth
google-auth-oauthlib

gunicorn==21.2.0
